[69] - Refactorizado StreamsDataProvider
