+++ tests/Unit/Services/StreamsDataManager/StreamsDataProviderTest.php
@@ -7,8 +7,8 @@ use App\Services\TokenProvider;
 use App\Config\TwitchConfig;
 use App\Infrastructure\Clients\APIClient;
 use App\Infrastructure\Serializers\StreamsDataSerializer;
+use Exception;
 use Illuminate\Http\Client\Response;
-use Illuminate\Support\Facades\Http;
 use Illuminate\Http\JsonResponse;
 use Tests\TestCase;
 use Mockery;
@@ -18,48 +18,38 @@ use Mockery;
  */
 class StreamsDataProviderTest extends TestCase
 {
-    public function testExecute()
+    /**
+     * @throws Exception
+     */
+    public function test_returns_json_response_with_serialized_data()
     {
-        // Mocks para las dependencias
-        $mockTokenProvider = Mockery::mock(TokenProvider::class);
-        $mockApiClient = Mockery::mock(APIClient::class);
-        $mockTwitchConfig = Mockery::mock(TwitchConfig::class);
-
-        // Configurar el comportamiento esperado de los mocks
-        $mockTokenProvider->shouldReceive('getToken')
+        $tokenProvider = Mockery::mock(TokenProvider::class);
+        $apiClient     = Mockery::mock(APIClient::class);
+        $twitchConfig  = Mockery::mock(TwitchConfig::class);
+        $tokenProvider->shouldReceive('getToken')
             ->once()
             ->andReturn('mocked_token');
-
-        $mockTwitchConfig->shouldReceive('clientId')
+        $twitchConfig->shouldReceive('clientId')
             ->once()
             ->andReturn('mocked_client_id');
-
         $expectedStreamsData = [
             'data' => [
                 ['title' => 'Stream 1', 'user_name' => 'user1'],
                 ['title' => 'Stream 2', 'user_name' => 'user2']
             ]
         ];
-
-        // Simular una respuesta HTTP exitosa
-        $mockResponse = Mockery::mock(Response::class);
-        $mockResponse->shouldReceive('successful')->andReturn(true);
-        $mockResponse->shouldReceive('json')->andReturn($expectedStreamsData);
-        $mockApiClient->shouldReceive('getDataForStreamsFromAPI')
+        $response = Mockery::mock(Response::class);
+        $response->shouldReceive('successful')->andReturn(true);
+        $response->shouldReceive('json')->andReturn($expectedStreamsData);
+        $apiClient->shouldReceive('getDataForStreamsFromAPI')
             ->with('mocked_client_id', 'mocked_token')
             ->once()
-            ->andReturn($mockResponse);
+            ->andReturn($response);
+        $streamsDataProvider = new StreamsDataProvider($tokenProvider, $apiClient, $twitchConfig);
 
-        // Instanciar la clase con los mocks
-        $provider = new StreamsDataProvider($mockTokenProvider, $mockApiClient, $mockTwitchConfig);
+        $response = $streamsDataProvider->execute();
 
-        // Ejecutar el método a testear
-        $response = $provider->execute();
-
-        // Serializar los datos como lo haría el método execute
         $serializedData = StreamsDataSerializer::serialize($expectedStreamsData['data']);
-
-        // Afirmar que la respuesta es correcta
         $this->assertInstanceOf(JsonResponse::class, $response);
         $this->assertEquals(200, $response->status());
         $this->assertEquals(json_encode($serializedData, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE), $response->content());
